Add bottom sheet recipe
This commit adds a recipe that demonstrates how to display a navigation
destination within a Material `ModalBottomSheet`.

It introduces `BottomSheetSceneStrategy`, a `SceneStrategy` that can be
added to a `NavDisplay`. This strategy checks for specific metadata on a
`NavEntry` to determine if it should be rendered as a bottom sheet.

A new `BottomSheetActivity` is added to showcase how to use this
strategy.