Add recipes for returning navigation results (#65)
Add recipes for returning navigation results
This commit adds two new recipes demonstrating how to return a result
from one screen to a previous one.

The "Return result as Event" recipe uses a `ResultEvent` API that allows passing results as event.

The "Return result as State" recipe with a `ResultStore` that can access the state as a Local.

Fixes: #34